Add migration option '-- +goose NO TRANSACTION' (#3)
The new option '-- +goose NO TRANSACTION' runs just that migration
outside of a transaction. This option applies to the entire file,
i.e. both 'goose up/down' sections. The default, of course, is to
use one independent DB transaction for each migration.

Migrations typically run inside transactions so that the migration
can be rolled back if there are problems with individual migration
statements or writing to goose metadata tables.

The use case to disable migration transactions is creating a postgres
index CONCURRENTLY, which can only be done outside a transaction.
See https://www.postgresql.org/docs/9.5/static/sql-createindex.html#SQL-CREATEINDEX-CONCURRENTLY

Also removed TestDriverSetFromEnvironmentVariable because it referenced
Sqlite3Dialect which is itself currently disabled. To re-add this test, replace
Sqlite3Dialect or rewrite this test.

A .travis.yml file was added to this repository, but may be incorrect and
automatic builds are not yet working.